Update README.md to warn users of AWS bills
This project walks the user through creating a Bedrock Knowledge base. This adds a warning that the Open Search Service has a baseline cost associated with it, so this cost is incurred by the user everyday until deletion! 
I learned this the hard way and so has atleast one other person who posted on reddit.